scxtop: address default-level clippy lints
Mechanical cleanup produced by cargo clippy --fix at default lint
levels, with manual review. No functional changes intended.

Lints addressed (all warn-by-default):
- clippy::collapsible_if
- clippy::let_and_return

+++ tools/scxtop/tests/test_wprof_annotations.rs
@@ -57,7 +57,7 @@ impl MockTrace {
                 annotations,
                 metadata: TrackEventMetadata {
                     cpu: Some((i % 8) as u32),
-                    pid: Some((1000 + i) as i32),
+                    pid: Some(((1000 + i))),
                     ..Default::default()
                 },
             });
